feature(core): split logic for fetching data and advanced filters
