+++ modules/@ergonode/channels/src/components/Tabs/ChannelHistoryTab.vue
@@ -20,7 +20,10 @@
 
 <script>
 import ResponsiveCenteredViewTemplate from '@Core/components/Layout/Templates/ResponsiveCenteredViewTemplate';
-import gridFetchDataMixin from '@Core/mixins/grid/gridFetchDataMixin';
+import {
+    DATA_LIMIT,
+} from '@Core/defaults/grid';
+import fetchGridDataMixin from '@Core/mixins/grid/fetchGridDataMixin';
 
 export default {
     name: 'ChannelHistoryTab',
@@ -28,10 +31,25 @@ export default {
         ResponsiveCenteredViewTemplate,
     },
     mixins: [
-        gridFetchDataMixin({
+        fetchGridDataMixin({
             path: 'channels/_id/exports',
         }),
     ],
+    fetch() {
+        return this.onFetchData({
+            offset: 0,
+            limit: DATA_LIMIT,
+            filters: '',
+            sortedColumn: {},
+        }).then(() => {
+            this.isPrefetchingData = false;
+        });
+    },
+    data() {
+        return {
+            isPrefetchingData: true,
+        };
+    },
     methods: {
         onEditRow(args) {
             const exportIndex = args.length - 1;
