+++ modules/@ergonode/collections/src/components/Tabs/CollectionGridTab.vue
@@ -23,7 +23,10 @@
 <script>
 import PRIVILEGES from '@Collections/config/privileges';
 import ResponsiveCenteredViewTemplate from '@Core/components/Layout/Templates/ResponsiveCenteredViewTemplate';
-import gridFetchDataMixin from '@Core/mixins/grid/gridFetchDataMixin';
+import {
+    DATA_LIMIT,
+} from '@Core/defaults/grid';
+import fetchGridDataMixin from '@Core/mixins/grid/fetchGridDataMixin';
 
 export default {
     name: 'CollectionGridTab',
@@ -31,10 +34,25 @@ export default {
         ResponsiveCenteredViewTemplate,
     },
     mixins: [
-        gridFetchDataMixin({
+        fetchGridDataMixin({
             path: 'collections',
         }),
     ],
+    fetch() {
+        return this.onFetchData({
+            offset: 0,
+            limit: DATA_LIMIT,
+            filters: '',
+            sortedColumn: {},
+        }).then(() => {
+            this.isPrefetchingData = false;
+        });
+    },
+    data() {
+        return {
+            isPrefetchingData: true,
+        };
+    },
     computed: {
         isUserAllowedToUpdate() {
             return this.$hasAccess([
