+++ modules/@ergonode/collections/src/components/Tabs/CollectionProductsTab.vue
@@ -52,10 +52,13 @@ import ActionButton from '@Core/components/ActionButton/ActionButton';
 import IconAdd from '@Core/components/Icons/Actions/IconAdd';
 import ResponsiveCenteredViewTemplate from '@Core/components/Layout/Templates/ResponsiveCenteredViewTemplate';
 import {
+    DATA_LIMIT,
+} from '@Core/defaults/grid';
+import {
     SIZE,
     THEME,
 } from '@Core/defaults/theme';
-import gridFetchDataMixin from '@Core/mixins/grid/gridFetchDataMixin';
+import fetchGridDataMixin from '@Core/mixins/grid/fetchGridDataMixin';
 import {
     mapState,
 } from 'vuex';
@@ -68,12 +71,23 @@ export default {
         IconAdd,
     },
     mixins: [
-        gridFetchDataMixin({
+        fetchGridDataMixin({
             path: 'collections/_id/elements',
         }),
     ],
+    fetch() {
+        return this.onFetchData({
+            offset: 0,
+            limit: DATA_LIMIT,
+            filters: '',
+            sortedColumn: {},
+        }).then(() => {
+            this.isPrefetchingData = false;
+        });
+    },
     data() {
         return {
+            isPrefetchingData: true,
             selectedAppModalOption: null,
         };
     },
