+++ modules/@ergonode/core/src/components/Tabs/UnitsSettingsGridTab.vue
@@ -40,10 +40,13 @@ import IconAdd from '@Core/components/Icons/Actions/IconAdd';
 import ResponsiveCenteredViewTemplate from '@Core/components/Layout/Templates/ResponsiveCenteredViewTemplate';
 import PRIVILEGES from '@Core/config/privileges';
 import {
+    DATA_LIMIT,
+} from '@Core/defaults/grid';
+import {
     SIZE,
     THEME,
 } from '@Core/defaults/theme';
-import gridFetchDataMixin from '@Core/mixins/grid/gridFetchDataMixin';
+import fetchGridDataMixin from '@Core/mixins/grid/fetchGridDataMixin';
 import {
     mapActions,
 } from 'vuex';
@@ -56,10 +59,25 @@ export default {
         IconAdd,
     },
     mixins: [
-        gridFetchDataMixin({
+        fetchGridDataMixin({
             path: 'units',
         }),
     ],
+    fetch() {
+        return this.onFetchData({
+            offset: 0,
+            limit: DATA_LIMIT,
+            filters: '',
+            sortedColumn: {},
+        }).then(() => {
+            this.isPrefetchingData = false;
+        });
+    },
+    data() {
+        return {
+            isPrefetchingData: true,
+        };
+    },
     computed: {
         smallSize() {
             return SIZE.SMALL;
