+++ modules/@ergonode/users/src/components/Tabs/RolesGridTab.vue
@@ -22,7 +22,10 @@
 
 <script>
 import ResponsiveCenteredViewTemplate from '@Core/components/Layout/Templates/ResponsiveCenteredViewTemplate';
-import gridFetchDataMixin from '@Core/mixins/grid/gridFetchDataMixin';
+import {
+    DATA_LIMIT,
+} from '@Core/defaults/grid';
+import fetchGridDataMixin from '@Core/mixins/grid/fetchGridDataMixin';
 import PRIVILEGES from '@Users/config/privileges';
 
 export default {
@@ -31,10 +34,25 @@ export default {
         ResponsiveCenteredViewTemplate,
     },
     mixins: [
-        gridFetchDataMixin({
+        fetchGridDataMixin({
             path: 'roles',
         }),
     ],
+    fetch() {
+        return this.onFetchData({
+            offset: 0,
+            limit: DATA_LIMIT,
+            filters: '',
+            sortedColumn: {},
+        }).then(() => {
+            this.isPrefetchingData = false;
+        });
+    },
+    data() {
+        return {
+            isPrefetchingData: true,
+        };
+    },
     computed: {
         isAllowedToUpdate() {
             return this.$hasAccess([
